pyproject.toml
[project]
name = "Discoin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "audioop-lts>=0.2.1",
    "discord-py>=2.5.2",
    "mistralai>=1.5.1",
    "python-dotenv>=1.0.1",
    "textual>=2.1.2",
    "beanie>=1.29.0",
    "coloredlogs>=15.0.1",
    "result>=0.17.0",
    "langchain>=0.3.20",
    "pypdf>=5.3.1",
    "validators>=0.34.0",
    "markdownify>=1.1.0",
    "sentry-sdk>=2.22.0",
]

[tool.setuptools]
py-modules = []
